fix path/move & check aabb

+++ src/main/java/com/nali/small/entities/skinning/ai/eyes/SkinningEntitiesLook.java
@@ -59,6 +59,7 @@ public class SkinningEntitiesLook extends SkinningEntitiesAI
         this.max = max;
         this.looking = true;
         this.done = false;
+//        this.onUpdate();//need update done
     }
 
     @Override
@@ -72,7 +73,7 @@ public class SkinningEntitiesLook extends SkinningEntitiesAI
 //
         if (this.looking)
         {
-            float yaw/*, pitch*/;
+            int yaw/*, pitch*/;
 
             this.looking = false;
             double d0 = this.x - this.skinningentities.posX;
@@ -88,7 +89,7 @@ public class SkinningEntitiesLook extends SkinningEntitiesAI
             if (riding_entity != null)
 //        for (Entity entity : this.skinningentities.getPassengers())
             {
-                yaw = riding_entity.rotationYaw;
+                yaw = (int)riding_entity.rotationYaw;
 //                pitch = riding_entity.rotationPitch;
                 riding_entity.rotationYaw = this.limitAngle(riding_entity.rotationYaw, f, this.max);
                 riding_entity.rotationPitch = this.limitAngle(riding_entity.rotationPitch, f1, this.max);
@@ -98,7 +99,7 @@ public class SkinningEntitiesLook extends SkinningEntitiesAI
 //                this.skinningentities.rotationYawHead = riding_entity.rotationYaw;
 
 //                if (yaw * yaw - riding_entity.rotationYaw * riding_entity.rotationYaw < 0.000001F && pitch * pitch - riding_entity.rotationPitch * riding_entity.rotationPitch < 0.000001F)
-                if (yaw == riding_entity.rotationYaw/* && pitch == riding_entity.rotationPitch*/)
+                if (yaw == (int)riding_entity.rotationYaw/* && pitch == riding_entity.rotationPitch*/)
 //                if (Math.abs(yaw - riding_entity.rotationYaw) < EPSILON && Math.abs(pitch - riding_entity.rotationPitch) < EPSILON)
                 {
                     this.done = true;
@@ -106,7 +107,7 @@ public class SkinningEntitiesLook extends SkinningEntitiesAI
             }
             else
             {
-                yaw = this.skinningentities.rotationYaw;
+                yaw = (int)this.skinningentities.rotationYaw;
 //                pitch = this.skinningentities.rotationPitch;
                 this.skinningentities.rotationYaw = this.limitAngle(this.skinningentities.rotationYaw, f, this.max);
                 this.skinningentities.rotationPitch = this.limitAngle(this.skinningentities.rotationPitch, f1, this.max);
@@ -118,7 +119,7 @@ public class SkinningEntitiesLook extends SkinningEntitiesAI
     //        this.skinningentities.prevRenderYawOffset = this.skinningentities.rotationYaw;
 
 //                if (yaw * yaw - this.skinningentities.renderYawOffset * this.skinningentities.renderYawOffset < 0.000001F && pitch * pitch - this.skinningentities.rotationPitch * this.skinningentities.rotationPitch < 0.000001F)
-                if (yaw == this.skinningentities.rotationYaw/* && pitch == this.skinningentities.rotationPitch*/)
+                if (yaw == (int)this.skinningentities.rotationYaw/* && pitch == this.skinningentities.rotationPitch*/)
 //                if (Math.abs(yaw - this.skinningentities.rotationYaw) < EPSILON && Math.abs(pitch - this.skinningentities.rotationPitch) < EPSILON)
                 {
                     this.done = true;
